RPOD.py changes
added a new method to the class description.

added a helper function graph_clusters which is
called in graph_jfh if use_clusters (which is set
to True in VisitingVehicle.py if
vv.set_cluster_config() is called in test_case_26.py
, otherwise it is set to False from the call of
vv.set_thruster_config() in test_case_26.py) is
True to create the clusterMesh and transform it
according to the ccf.

graph_jfh: added call to graph_clusters if
use_clusters is True. added the required
translation of the plumeMesh (also functionally
operating as the mesh of the thruster) with
respect to its corresponding cluster position.
concatenated the active_clusters.data to the
VVmesh if use_clusters is True.

jfh_plume_strikes: adjusted the plume strike
location accordingly if use_clusters is True